Refactored Valuation component to streamline state management and improve search functionality. Removed unused states and API calls, integrated CarValuationForm and ValuationBar components for better UI structure, and updated result handling logic.
